+++ src/synthesizer/abundances/__init__.py
@@ -1,0 +1,3 @@
+from synthesizer.abundances.abundances import (
+    Abundances
+)
